+++ internal/test-utils-core/src/commonMain/kotlin/com/github/panpf/sketch/test/utils/decodes.common.kt
@@ -2,6 +2,7 @@ package com.github.panpf.sketch.test.utils
 
 import com.github.panpf.sketch.Image
 import com.github.panpf.sketch.Sketch
+import com.github.panpf.sketch.decode.BitmapConfig
 import com.github.panpf.sketch.decode.DecodeResult
 import com.github.panpf.sketch.decode.Decoder
 import com.github.panpf.sketch.decode.internal.DecodeHelper
@@ -21,9 +22,11 @@ suspend fun ImageRequest.decode(sketch: Sketch, factory: Decoder.Factory? = null
 
 expect fun createDecodeHelper(request: ImageRequest, dataSource: DataSource): DecodeHelper
 
-fun ResourceImageFile.decode(): Image {
+fun ResourceImageFile.decode(bitmapConfig: BitmapConfig? = null): Image {
     val context = getTestContext()
-    val request = ImageRequest(context, uri)
+    val request = ImageRequest(context, uri) {
+        bitmapConfig(bitmapConfig)
+    }
     val dataSource = toDataSource(context)
     val decoderHelper = createDecodeHelper(request, dataSource)
     return decoderHelper.decode(1)