+++ sample/src/androidMain/kotlin/com/github/panpf/sketch/sample/ui/test/transform/RotateTransformationTestFragment.kt
@@ -17,6 +17,7 @@
 package com.github.panpf.sketch.sample.ui.test.transform
 
 import android.app.Application
+import android.graphics.Bitmap
 import android.os.Bundle
 import android.widget.SeekBar
 import androidx.fragment.app.viewModels
@@ -24,6 +25,7 @@ import androidx.lifecycle.Lifecycle.State
 import com.github.panpf.sketch.cache.CachePolicy.DISABLED
 import com.github.panpf.sketch.images.ResourceImages
 import com.github.panpf.sketch.loadImage
+import com.github.panpf.sketch.request.bitmapConfig
 import com.github.panpf.sketch.sample.databinding.FragmentTestTransformationRotateBinding
 import com.github.panpf.sketch.sample.ui.base.BaseBindingFragment
 import com.github.panpf.sketch.sample.ui.base.LifecycleAndroidViewModel
@@ -46,6 +48,7 @@ class RotateTransformationTestFragment :
                 memoryCachePolicy(DISABLED)
                 resultCachePolicy(DISABLED)
                 addTransformations(RotateTransformation(it))
+                bitmapConfig(Bitmap.Config.RGB_565) // To test automatic conversion Config
             }
 
             binding.degreesText.text = "$it"