+++ sample/src/commonMain/kotlin/com/github/panpf/sketch/sample/ui/test/transform/BlurTransformationTestPage.kt
@@ -22,10 +22,10 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.github.panpf.sketch.cache.CachePolicy.DISABLED
 import com.github.panpf.sketch.LocalPlatformContext
-import com.github.panpf.sketch.request.ImageRequest
+import com.github.panpf.sketch.cache.CachePolicy.DISABLED
 import com.github.panpf.sketch.images.ResourceImages
+import com.github.panpf.sketch.request.ImageRequest
 import com.github.panpf.sketch.sample.ui.components.MyAsyncImage
 import com.github.panpf.sketch.transform.BlurTransformation
 import kotlin.math.roundToInt
@@ -59,6 +59,7 @@ fun BlurTransformationTestPage() {
                             maskColor = maskColor
                         )
                     )
+//                    bitmapConfig(BitmapConfig.FixedQuality("RGB_565")) // TODO There is a problem
                 },
                 contentDescription = "image",
                 modifier = Modifier.fillMaxHeight().weight(1f)