+++ sample/src/commonMain/kotlin/com/github/panpf/sketch/sample/ui/test/transform/CircleCropTransformationTestPage.kt
@@ -18,11 +18,12 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.github.panpf.sketch.cache.CachePolicy.DISABLED
 import com.github.panpf.sketch.LocalPlatformContext
+import com.github.panpf.sketch.cache.CachePolicy.DISABLED
+import com.github.panpf.sketch.decode.BitmapConfig
+import com.github.panpf.sketch.images.ResourceImages
 import com.github.panpf.sketch.request.ImageRequest
 import com.github.panpf.sketch.resize.Scale
-import com.github.panpf.sketch.images.ResourceImages
 import com.github.panpf.sketch.sample.ui.components.MyAsyncImage
 import com.github.panpf.sketch.transform.CircleCropTransformation
 
@@ -35,6 +36,7 @@ fun CircleTransformationTestPage() {
                 memoryCachePolicy(DISABLED)
                 resultCachePolicy(DISABLED)
                 addTransformations(CircleCropTransformation(scale))
+                bitmapConfig(BitmapConfig.FixedQuality("RGB_565")) // To test automatic conversion Config
             },
             contentDescription = "image",
             modifier = Modifier.fillMaxWidth().weight(1f)